Adding a new SOR code will no longer automatically scroll the page upwards. Moved NewSorCode component into Components folder
